chore: rename common -> cosmic_text
